Fixed the way that trees were setting depths
The way that trees were setting weights was not taking into account the
final branch, which needs to be set such that the tree is ultrametric.
This wasn't being done.

One thing to note, the unroot might not be getting taken into account,
so the depths will be off by one layer. I didn't carefully check before
commiting.