Investigated and documented a instability based on order of trees
There is an instability associated with NJ and the order in which trees are
presented. Two things to know

-   NJ in unstable due to ties
-   The order of taxa in a distance table is arbitrary

Therefore, while NJ resolves its ties consistently, (it always takes the latest
pair), the order those ties are presented are not always consistent. The order
they are presented is based on the first tree given to SunStar. Therefore,
we have an instability based on order.

Probably the best way to resolve this is to have NJ flip a coin when it finds a
tie. This will make testing hard. I might accept if its an odd loop iteration,
but I'm not sure that is much more fair. Regardless, we need to "distribute"
the choices that NJ makes.