Changed some error conditions in the newick parser, commented.
There were still assert false's in this code to "handle" errors. The old logic
was that exceptions are kinda expensive, so I shouldn't use them. The problem
with this is that the initial parsing of the newick trees is something like
O(nm) where n is the number of taxa on a tree, and m is the number of trees.
This is pretty small in comparision to the amount of time spent on the real
computation. So, I am switching to using exceptions with the idea that I can
better clue the user into handling errors relating to file input.