Made the newick grammar more permissive w.r.t. labels
Originally, when I had written the newick parser, I had thought to make
the grammar as simple as possible for me, the writer. Now that I have
stepped away from it, I don't want to deal with converting other peoples
labels to my style, so I just made it so that only a few characters are
disallowed.